Move region CSV code to `input/region.rs`
