+++ src/input.rs
@@ -7,6 +7,9 @@ use std::fs;
 use std::path::Path;
 use std::rc::Rc;
 
+pub mod region;
+pub use region::read_regions;
+
 /// Read a series of type `T`s from a CSV file.
 ///
 /// # Arguments
